Fixed issue #52
Former-commit-id: 7f7717811396c680aeae2105e38a8196e576227c

+++ src/main/java/fr/inria/cedar/quotientSummary/util/RDF2SQLEncoding.java
@@ -133,7 +133,7 @@ public class RDF2SQLEncoding {
 			try (ResultSet rs = stmtEncode.executeQuery()) {
 				//LOGGER.debug("Asked query: " + learnCodeQueryString);
 				if (rs.next()) {
-					code = rs.getInt(1);
+					code = rs.getLong(1);
 					//LOGGER.debug("The code of " + URI + " is: " + constantCode);
 				}
 			}
